Move source code to a src folder for easier management.
