+++ README.md
@@ -10,3 +10,5 @@ Example golang webservice that is dockerized, and easily loads up in to GoLand.
 
 * [Design](docs/design.md)
 * [Development](docs/development.md)
+* [SRE](docs/sre.md)
+* [Testing](docs/testing.md)
