--- admin/go.mod
@@ -1,3 +1,0 @@
-module github.com/clickandobey/golang-dockerized-webservice/admin
-
-go 1.15
