+++ docs/sre.md
@@ -1,0 +1,3 @@
+# Site Reliability Engineering
+
+It is suggested you use the SRE Checklist and other docs found [here](https://github.com/ClickAndObey/sre-checklist).
