+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/clickandobey/golang-dockerized-webservice/webservice"
+import "github.com/clickandobey/golang-dockerized-webservice/src/webservice"
 
 func main() {
 	webservice.Run()
