+++ src/admin/configuration_test.go
