+++ src/admin/go.mod
@@ -1,0 +1,3 @@
+module github.com/clickandobey/golang-dockerized-webservice/src/admin
+
+go 1.15
