+++ src/go.mod
@@ -1,0 +1,3 @@
+module github.com/clickandobey/golang-dockerized-webservice/src
+
+go 1.15
